remove fragile tests
